Created PhotoSqlDAO

+++ java/src/main/java/com/techelevator/dao/UserSqlDAO.java
@@ -28,7 +28,7 @@ public class UserSqlDAO implements UserDAO {
     }
 
 	@Override
-	public User getUserById(Long userId) {
+	public User getUserById(int userId) {
 		String sql = "SELECT * FROM users WHERE user_id = ?";
 		SqlRowSet results = jdbcTemplate.queryForRowSet(sql, userId);
 		if(results.next()) {
@@ -88,7 +88,7 @@ public class UserSqlDAO implements UserDAO {
 
     private User mapRowToUser(SqlRowSet rs) {
         User user = new User();
-        user.setId(rs.getLong("user_id"));
+        user.setId(rs.getInt("user_id"));
         user.setUsername(rs.getString("username"));
         user.setPassword(rs.getString("password_hash"));
         user.setAuthorities(rs.getString("role"));
